+++ src/primaite/simulator/network/hardware/base.py
@@ -547,15 +547,15 @@ class Link(SimComponent):
 
     :param endpoint_a: The first NIC or SwitchPort connected to the Link.
     :param endpoint_b: The second NIC or SwitchPort connected to the Link.
-    :param bandwidth: The bandwidth of the Link in Mbps (default is 100 Mbps).
+    :param bandwidth: The bandwidth of the Link in Mbps.
     """
 
     endpoint_a: WiredNetworkInterface
     "The first WiredNetworkInterface connected to the Link."
     endpoint_b: WiredNetworkInterface
     "The second WiredNetworkInterface connected to the Link."
-    bandwidth: float = 100.0
-    "The bandwidth of the Link in Mbps (default is 100 Mbps)."
+    bandwidth: float
+    "The bandwidth of the Link in Mbps."
     current_load: float = 0.0
     "The current load on the link in Mbps."